requirements.txt
panel
python-dotenv
langchain-openai
langchain-groq
langchain-google-genai
langchain-community
faiss-cpu
chromadb
beautifulsoup4
